VideoRecog/requirements.txt
watchdog
python-socketio[client]